+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this provider will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+* Added arn column for security group resource [#278](https://github.com/cloudquery/cq-provider-aws/issues/277)
+
+
 ## [v0.6.4] - 2021-11-16
 ###### SDK Version: 0.5.1